backend/requirements.txt
# Django Backend Requirements
# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Machine Learning
tensorflow==2.19.0
scikit-learn==1.6.1
numpy==2.0.2
pandas==2.2.2
tensorflow-hub==0.16.1
joblib==1.5.1

# Audio Processing
librosa==0.11.0
soundfile==0.13.1
pydub==0.25.1

# API and File Handling
Pillow>=9.5.0,<10.0.0
python-multipart==0.0.6

# Environment and Config
python-dotenv==1.0.0
PyYAML==6.0.1

# Development Tools
whitenoise==6.6.0

# Database - MySQL
PyMySQL==1.1.0

# Utilities
requests==2.31.0

# Visualization
matplotlib>=3.10.0
tqdm==4.67.1

# Model explainabilit

# Authentication
djangorestframework-simplejwt==5.3.1